requirements_dev.txt
pip==19.2.3
bump2version==0.5.11
wheel==0.33.6
watchdog==0.9.0
tox==3.14.0
coverage==4.5.4
Sphinx==7.2.6
twine==5.0.0
ruff==0.3.5
llama-index-core
llama-index-legacy
llama-index-llms-openai
llama-index-embeddings-openai
llama-index-program-openai
llama-index-question-gen-openai
llama-index-agent-openai
llama-index-readers-file
llama-index-multi-modal-llms-openai
qdrant-client
fastembed
sentence_transformers
python-dotenv